pyproject.toml
[project]
name = "voting-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eth-ape>=0.8.36",
    "ruff>=0.11.13",
]
